Add reading shops HTML selectors from XML file. Add moving shops XML config file to UT and main plugins file.

+++ Sources/UnitTest/PluginsTests/CProlineParserTest.cpp
@@ -106,7 +106,7 @@ void CProlineParserTest::testPluginParsing_ProductPage()
 	QVERIFY(!stProlineTest.isLoadingNextSearchPage());
 
 	stProlineTest.parseProductPage();
-	QCOMPARE(double(stProlineTest.m_stCompData.dPrice),double(579));
+	QCOMPARE(double(stProlineTest.m_stCompData.dPrice),double(355));
 	QCOMPARE(stProlineTest.m_stCompData.eSearchResult,SR_COMPFOUND);
 }
 void CProlineParserTest::testPluginParsing_NotFoundPage()
